# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: HAPE API
  version: 1.0.0
paths:
...
  /auth/tokens:
    post:
      tags:
      - auth
      summary: Create Token
      operationId: create_token_auth_tokens_post
      parameters:
      - name: x-hape-admin-key
        in: header
        required: false
        schema:
          type: string
          default: ''
          title: X-Hape-Admin-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTokenRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateTokenResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - auth
      summary: List Tokens
      operationId: list_tokens_auth_tokens_get
      parameters:
      - name: x-hape-admin-key
        in: header
        required: false
        schema:
          type: string
          default: ''
          title: X-Hape-Admin-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TokenMetadata'
                title: Response List Tokens Auth Tokens Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /auth/tokens/revoke:
    post:
      tags:
      - auth
      summary: Revoke Token
      operationId: revoke_token_auth_tokens_revoke_post
      parameters:
      - name: x-hape-admin-key
        in: header
        required: false
        schema:
          type: string
          default: ''
          title: X-Hape-Admin-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RevokeTokenRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
                title: Response Revoke Token Auth Tokens Revoke Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CreateTokenRequest:
      properties:
        name:
          type: string
          title: Name
      type: object
      required:
      - name
      title: CreateTokenRequest
    CreateTokenResponse:
      properties:
        token:
          type: string
          title: Token
        token_id:
          type: string
          title: Token Id
        name:
          type: string
          title: Name
        created_at:
          type: string
          title: Created At
      type: object
      required:
      - token
      - token_id
      - name
      - created_at
      title: CreateTokenResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    RevokeTokenRequest:
      properties:
        token_id:
          type: string
          title: Token Id
      type: object
      required:
      - token_id
      title: RevokeTokenRequest
...
    TokenMetadata:
      properties:
        token_id:
          type: string
          title: Token Id
        name:
          type: string
          title: Name
        created_at:
          type: string
          title: Created At
        revoked:
          type: boolean
          title: Revoked
      type: object
      required:
      - token_id
      - name
      - created_at
      - revoked
      title: TokenMetadata
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError